Convert to OpenGL rendering.
This now renders live views of several of the fractal generators. Currently interactive mandelbrot and non-interactive DeJong Attractors work. They render to an OpenGL texture which is then rendered over the entire screen. DeJong even shows off rendering in a background thread. I have plans to move mandelbrot to a sequential rendering so progress can be seen, but thats still in progress. Also in progress are other generators like Perlin and DiamondSquare.